Create two-level dropdown instead of recursive solution.
